src/requirements.txt
uvicorn==0.29.0
fastapi-sso==0.15.0
fastapi==0.110.1
python-jose[cryptography]==3.3.0
jinja2==3.1.3
sqlmodel==0.0.16
azure-storage-blob==12.19.1
python-multipart==0.0.9
aiohttp==3.9.5
azure-monitor-opentelemetry==1.6.0